+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 - update @motionpicture/coa-service
 - COA予約承認処理をMGチケットに対応
+- 口座決済返金時のtoocationのタイプをAccountに変更
 
 ### Deprecated